% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-aggregate.R
\name{aggregate_sentiment}
\alias{aggregate_sentiment}
\title{Get statistics for sentiment over some group, such as sentence.}
\usage{
aggregate_sentiment(.data, aggregate_on, statistic)
}
\arguments{
\item{.data}{character vector of words}

\item{aggregate_on}{vector to aggregate .data over; ideally,
sentence_id, but could be chapter, document, etc.}

\item{statistic}{function that accepts na.rm argument; e.g. mean,
median, sd.}
}
\description{
Get statistics for sentiment over some group, such as sentence.
}
