% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-aggregate.R
\name{bind_aggregation}
\alias{bind_aggregation}
\title{bind aggregate terms together}
\usage{
bind_aggregation(data, aggregate_on)
}
\arguments{
\item{data}{vector of terms}

\item{aggregate_on}{vector of aggregations}
}
\value{
data with every aggregation bound, as in a sentence
}
\description{
bind aggregate terms together
}
