% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-term.R
\name{concat_walk}
\alias{concat_walk}
\title{concat list 1 and 2, moving past NA values}
\usage{
concat_walk(list1, list2)
}
\arguments{
\item{list1}{list or vector for first bigram token}

\item{list2}{list or vector for second bigram token}
}
\value{
paste of list1 and list2, skipping NA's
}
\description{
concat list 1 and 2, moving past NA values
}
