% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-term.R
\name{concat_walk_i}
\alias{concat_walk_i}
\title{concat list 1 and 2 at index, skipping NA values}
\usage{
concat_walk_i(i, list1, list2)
}
\arguments{
\item{i}{numeric index to assess index at}

\item{list1}{list or vector for first token}

\item{list2}{list or vector for second token}
}
\value{
paste of list1 and list2 at index i, skipping NA's
}
\description{
concat list 1 and 2 at index, skipping NA values
}
