% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation-format.R
\name{determine_stopwords}
\alias{determine_stopwords}
\title{determine stopword status}
\usage{
determine_stopwords(.data, ...)
}
\arguments{
\item{.data}{vector of words}

\item{...}{arguments of get_sw}
}
\value{
a [tibble][tibble::tibble-package] equivalent to the input
  dataframe, with an additional stopword column
}
\description{
determine stopword status
}
