% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dist_density}
\alias{dist_density}
\title{output a histogram of the distribution of some function of words}
\usage{
dist_density(.data, col_name)
}
\arguments{
\item{.data}{the standard dataframe, modified so the last column
is the output of some insight function (eg. output from
term_freq)}

\item{col_name}{symbol name of the column insight was
performed on}
}
\description{
output a histogram of the distribution of some function of words
}
