% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation-format.R
\name{format_data}
\alias{format_data}
\title{formats imported data into an analysis-ready format '}
\usage{
format_data(data)
}
\arguments{
\item{data}{a tibble formatted with a text and (optional) group
column}
}
\value{
a [tibble][tibble::tibble-package] formatted such that
    columns correspond to identifiers of group, line, sentence,
    word (groups ignored)
}
\description{
formats imported data into an analysis-ready format '
}
