% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-aggregate.R
\name{get_aggregate_insight}
\alias{get_aggregate_insight}
\title{perform group-aware aggregate operations on the data}
\usage{
get_aggregate_insight(.data, operations, aggregate_on, ...)
}
\arguments{
\item{.data}{dataframe of terms as per output of text_prep}

\item{operations}{character vector of operations to perform}

\item{aggregate_on}{character name of the column to perform aggregate operations on}

\item{...}{additional arguments to the operation - only sensible for singular operations}
}
\value{
.data with operation columns added
}
\description{
perform group-aware aggregate operations on the data
}
