% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-term.R
\name{get_bigram}
\alias{get_bigram}
\title{Determine bigrams}
\usage{
get_bigram(.data)
}
\arguments{
\item{.data}{character vector of words}
}
\value{
character vector of bigrams
}
\description{
Determine bigrams
}
