% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation-format.R
\name{get_books}
\alias{get_books}
\title{sections text based on book}
\usage{
get_books(.data)
}
\arguments{
\item{.data}{vector to section}
}
\value{
vector of same length as .data with book numbers
}
\description{
sections text based on book
}
