% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{get_cantos}
\alias{get_cantos}
\title{sections text based on cantos}
\usage{
get_cantos(.data)
}
\arguments{
\item{.data}{vector to section}
}
\value{
vector of same length as .data with canto numbers
}
\description{
sections text based on cantos
}
