% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{get_chapters}
\alias{get_chapters}
\title{sections text based on chapters}
\usage{
get_chapters(.data)
}
\arguments{
\item{.data}{vector to section}
}
\value{
vector of same length as .data with chapter numbers
}
\description{
sections text based on chapters
}
