% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_filetype}
\alias{get_filetype}
\title{Get filetype}
\usage{
get_filetype(filepath)
}
\arguments{
\item{filepath}{string filepath of document}
}
\value{
filetype (string) - NA if no extension
}
\description{
Get filetype
}
