% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-term.R
\name{get_ngram}
\alias{get_ngram}
\title{Returns the n-grams, skipping NA values}
\usage{
get_ngram(.data, n)
}
\arguments{
\item{.data}{vector to get n-grams from}

\item{n}{number of n-grams to attain}
}
\value{
n-gram vector without NA values
}
\description{
Returns the n-grams, skipping NA values
}
