% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{get_parts}
\alias{get_parts}
\title{sections text based on parts}
\usage{
get_parts(.data)
}
\arguments{
\item{.data}{vector to section}
}
\value{
vector of same length as .data with part numbers
}
\description{
sections text based on parts
}
