% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{get_search}
\alias{get_search}
\title{creates a search closure to section text}
\usage{
get_search(search)
}
\arguments{
\item{search}{a string regexp for the term to seperate on, e.g. "Chapter"}
}
\value{
closure over search expression
}
\description{
creates a search closure to section text
}
