% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation-format.R
\name{get_sections}
\alias{get_sections}
\title{sections text based on sections}
\usage{
get_sections(.data)
}
\arguments{
\item{.data}{vector to section}
}
\value{
vector of same length as .data with section numbers
}
\description{
sections text based on sections
}
