% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{get_sw}
\alias{get_sw}
\title{Gets stopwords from a default list and user-provided list}
\usage{
get_sw(lexicon = "snowball", addl = NA)
}
\arguments{
\item{lexicon}{a string name of a stopword list, one of "smart",
"snowball", or "onix"}

\item{addl}{user defined character vector of additional stopwords,
each element being a stopword}
}
\value{
a [tibble][tibble::tibble-package] with one column named "word"
}
\description{
Gets stopwords from a default list and user-provided list
}
