% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-aggregate.R
\name{get_term_insight}
\alias{get_term_insight}
\title{perform group-aware term operations on the data}
\usage{
get_term_insight(.data, operations)
}
\arguments{
\item{.data}{dataframe of terms as per output of text_prep}

\item{operations}{character vector of term operations to perform}
}
\value{
.data with operation columns added
}
\description{
perform group-aware term operations on the data
}
