% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_valid_input}
\alias{get_valid_input}
\title{helper function to get valid input (recursively)}
\usage{
get_valid_input(options, init = TRUE)
}
\arguments{
\item{options}{vector of options that valid input should be drawn from}

\item{init}{whether this is the initial attempt, used only as
recursive information}
}
\value{
readline output that exists in the vector of options
}
\description{
helper function to get valid input (recursively)
}
