% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-vis.R
\name{get_vis}
\alias{get_vis}
\title{create a group-aware visualisation}
\usage{
get_vis(.data, vis, col, facet_by = "", scale_fixed = TRUE, ncol = 3, ...)
}
\arguments{
\item{.data}{the standard dataframe, modified so the last column
is the output of some insight function (eg. output from
term_freq)}

\item{vis}{character name of visualisation function}

\item{col}{character name of the column to get insight from}

\item{facet_by}{character name of the column to facet by}

\item{scale_fixed}{force scales to be fixed in a facet}

\item{ncol}{number of columns for facet wrap}

\item{...}{additional arguments to the visualisation}
}
\description{
create a group-aware visualisation
}
