% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ifexp}
\alias{ifexp}
\title{scheme-like if expression, without restriction of returning same-size table of .test, as ifelse() does}
\usage{
ifexp(.test, true, false)
}
\arguments{
\item{.test}{predicate to test}

\item{true}{expression to return if .test evals to TRUE}

\item{false}{expression to return if .test evals to TRUE}
}
\value{
either true or false
}
\description{
scheme-like if expression, without restriction of returning same-size table of .test, as ifelse() does
}
