% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-files.R
\name{import_base_file}
\alias{import_base_file}
\title{Base case for file import}
\usage{
import_base_file(filepath)
}
\arguments{
\item{filepath}{string filepath of file for import}
}
\value{
imported file with document id
}
\description{
Base case for file import
}
