% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-csv.R
\name{import_csv}
\alias{import_csv}
\title{Import csv file}
\usage{
import_csv(filepath)
}
\arguments{
\item{filepath}{a string indicating the relative or absolute
filepath of the file to import}
}
\value{
a [tibble][tibble::tibble-package] of each row
  corrresponding to a line of the text file, with the column named
  "text"
}
\description{
Import csv file
}
