% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation-import.R
\name{import_files}
\alias{import_files}
\title{Import any number of files}
\usage{
import_files(filepaths)
}
\arguments{
\item{filepaths}{char vector of filepaths}
}
\value{
imported files with document id, bound into tibble
}
\description{
Import any number of files
}
