% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-term.R
\name{index_bigram}
\alias{index_bigram}
\title{get bigram at index i of list1 & 2}
\usage{
index_bigram(i, list1, list2)
}
\arguments{
\item{i}{numeric index to attain bigram at}

\item{list1}{list or vector for first bigram token}

\item{list2}{list or vector for second bigram token}
}
\value{
bigram of list1 and list2 at index i, skipping NA's
}
\description{
get bigram at index i of list1 & 2
}
