% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-aggregate.R
\name{key_sentences}
\alias{key_sentences}
\title{get score for key sentences as per Lexrank}
\usage{
key_sentences(.data, aggregate_on)
}
\arguments{
\item{.data}{character vector of words}

\item{aggregate_on}{vector to aggregate .data over; ideally, sentence_id}
}
\description{
get score for key sentences as per Lexrank
}
