% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-term.R
\name{keywords_tr}
\alias{keywords_tr}
\title{Determine textrank score for vector of words}
\usage{
keywords_tr(.data)
}
\arguments{
\item{.data}{character vector of words}
}
\value{
vector of scores for each word
}
\description{
Determine textrank score for vector of words
}
