% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma-term-sentiment.R
\name{ma_term_sentiment}
\alias{ma_term_sentiment}
\title{Determine the lagged sentiment of terms}
\usage{
ma_term_sentiment(.data, lexicon = "afinn", lag = 10, statistic = mean)
}
\arguments{
\item{.data}{vector of terms}

\item{lexicon}{sentiment lexicon to use, based on the corpus
provided by tidytext}

\item{lag}{how many (inclusive) terms to compute statistic over}

\item{statistic}{base statistic used to summarise the data, capable
of taking an na.rm argument}
}
\value{
vector with lagged sentiment score of each term in the input vector
}
\description{
Determine the lagged sentiment of terms
}
