% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram-freq.R
\name{ngram_freq}
\alias{ngram_freq}
\title{NOT FOR PRODUCTION - STILL IN TESING. Returns the count of n-grams, skipping NA values}
\usage{
ngram_freq(.data, n)
}
\arguments{
\item{.data}{vector to get n-grams from}

\item{n}{number of n-grams to attain}
}
\value{
count of each associated n-gram
}
\description{
NOT FOR PRODUCTION - STILL IN TESING. Returns the count of n-grams, skipping NA values
}
