% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{score_bar_ungrouped}
\alias{score_bar_ungrouped}
\title{output a ggplot column graph of the top texts from some insight function}
\usage{
score_bar_ungrouped(.data, y, n = 15, x = text, desc = FALSE)
}
\arguments{
\item{.data}{a dataframe containing "text" and insight columns as
per the output of the get_(term|aggregate)_insight wrapper
function}

\item{y}{symbol name of the column insight was
outputted to}

\item{n}{number of bars to display}

\item{x}{symbol name of column for insight labels}

\item{desc}{bool: show bars in descending order}
}
\description{
output a ggplot column graph of the top texts from some insight function
}
