% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.R
\name{score_wordcloud}
\alias{score_wordcloud}
\title{output a ggplot wordcloud graph of the top texts from some insight function}
\usage{
score_wordcloud(.data, y, n = 15, x = text, shape = "circle")
}
\arguments{
\item{.data}{a dataframe containing "text" and insight columns as
per the output of the get_(term|aggregate)_insight wrapper
function}

\item{y}{symbol name of the column insight was
outputted to}

\item{n}{number of words to display}

\item{x}{symbol name of column for insight labels}

\item{shape}{character: shape of the wordcloud}
}
\description{
output a ggplot wordcloud graph of the top texts from some insight function
}
