% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{section}
\alias{section}
\title{Adds section column to dataframe}
\usage{
section(.data, section_by)
}
\arguments{
\item{.data}{dataframe formatted as per output of prep process}

\item{section_by}{character name of what to section over}
}
\value{
input dataframe with additional section column
}
\description{
Adds section column to dataframe
}
