% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{shorten}
\alias{shorten}
\title{Shorten some text up to n characters}
\usage{
shorten(.data, n)
}
\arguments{
\item{.data}{character vector}

\item{n}{wrap length of text}
}
\value{
shortened form of .data
}
\description{
Shorten some text up to n characters
}
