% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{struct_ggpage_ungrouped}
\alias{struct_ggpage_ungrouped}
\title{Colours a ggpage based on an insight function}
\usage{
struct_ggpage_ungrouped(.data, col_name)
}
\arguments{
\item{.data}{a dataframe containing "word" and insight columns as
per the output of the get_(term|aggregate)_insight wrapper
function}

\item{col_name}{symbol name of the insight column intended to
colour plot}
}
\value{
ggplot object as per ggpage
}
\description{
Colours a ggpage based on an insight function
}
