% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{struct_ts_ungrouped}
\alias{struct_ts_ungrouped}
\title{output a ggplot time series plot of some insight function}
\usage{
struct_ts_ungrouped(.data, y)
}
\arguments{
\item{.data}{a dataframe containing "text" and insight columns as
per the output of the get_(term|aggregate)_insight wrapper
function}

\item{y}{symbol name of the column insight was
outputted to}
}
\description{
output a ggplot time series plot of some insight function
}
