% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_textcol}
\alias{table_textcol}
\title{Interactively determine and automatically mark the text column of a table}
\usage{
table_textcol(data)
}
\arguments{
\item{data}{dataframe with column requiring marking}
}
\value{
same dataframe with text column renamed to "text"
}
\description{
Interactively determine and automatically mark the text column of a table
}
