% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-term.R
\name{term_corr}
\alias{term_corr}
\title{Determine term correlations - extremely slow}
\usage{
term_corr(.data, term, aggregate_on)
}
\arguments{
\item{.data}{character vector of terms}

\item{term}{character to find correlations with}

\item{aggregate_on}{vector to aggregate .data over; ideally,
sentence_id, but could be chapter, document, etc.}
}
\value{
numeric vector of term correlations as per phi_coef
}
\description{
Determine term correlations - extremely slow
}
