% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-aggregate.R
\name{term_count}
\alias{term_count}
\title{Determine the number of terms at each aggregate level}
\usage{
term_count(.data, aggregate_on)
}
\arguments{
\item{.data}{character vector of terms}

\item{aggregate_on}{vector to split .data on for insight}
}
\value{
vector of number of terms for each aggregate level, same
  length as .data
}
\description{
Determine the number of terms at each aggregate level
}
