% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insight-term.R
\name{term_freq}
\alias{term_freq}
\title{Determine term frequency}
\usage{
term_freq(.data)
}
\arguments{
\item{.data}{character vector of terms}
}
\value{
numeric vector of term frequencies
}
\description{
Determine term frequency
}
