% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation-format.R
\name{text_prep}
\alias{text_prep}
\title{takes imported one-line-per-row data and prepares it for later analysis}
\usage{
text_prep(.data, lemmatize = TRUE, stopwords = TRUE,
  sw_lexicon = "snowball", addl_stopwords = NA)
}
\arguments{
\item{.data}{tibble with one line of text per row}

\item{lemmatize}{boolean, whether to lemmatize or not}

\item{stopwords}{boolean, whether to remove stopwords or not}

\item{sw_lexicon}{string, lexicon with which to remove stopwords}

\item{addl_stopwords}{char vector of user-supplied stopwords}
}
\value{
tibble with one token per line, stopwords removed leaving
  NA values, column for analysis named "text"
}
\description{
takes imported one-line-per-row data and prepares it for later analysis
}
