% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ungroup_by}
\alias{ungroup_by}
\title{helper function to ungroup for dplyr. functions equivalently to
group_by() but with standard (string) evaluation}
\usage{
ungroup_by(x, ...)
}
\arguments{
\item{x}{tibble to perform function on}

\item{...}{string of groups to ungroup on}
}
\value{
x with ... no longer grouped upon
}
\description{
helper function to ungroup for dplyr. functions equivalently to
group_by() but with standard (string) evaluation
}
