% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-data.R
\name{format_data}
\alias{format_data}
\title{takes imported one-line-per-row data and prepares it for later analysis}
\usage{
format_data(
  .data,
  lemmatize = TRUE,
  stopwords = TRUE,
  sw_lexicon = "snowball",
  addl_stopwords = NA
)
}
\arguments{
\item{.data}{tibble with one line of text per row}

\item{lemmatize}{boolean, whether to lemmatize or not}

\item{stopwords}{boolean, whether to remove stopwords or not}

\item{sw_lexicon}{string, lexicon with which to remove stopwords}

\item{addl_stopwords}{char vector of user-supplied stopwords}
}
\value{
a [tibble][tibble::tibble-package] with one token per line,
  stopwords removed leaving NA values, column for analysis named
  "text"
}
\description{
takes imported one-line-per-row data and prepares it for later analysis
}
