% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-files.R
\name{import_files}
\alias{import_files}
\title{Import any number of files}
\usage{
import_files(filepaths, filenames)
}
\arguments{
\item{filepaths}{char vector of filepaths}

\item{filenames}{char vector of file names}
}
\value{
a [tibble][tibble::tibble-package] imported files with
  document id
}
\description{
Import any number of files
}
