% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-sentiment.R
\name{term_sentiment}
\alias{term_sentiment}
\title{Determine sentiment of terms}
\usage{
term_sentiment(.data, lexicon = "afinn")
}
\arguments{
\item{.data}{vector of terms}

\item{lexicon}{sentiment lexicon to use, based on the corpus
provided by tidytext}
}
\value{
vector with sentiment score of each word in the vector
}
\description{
Determine sentiment of terms
}
